Splits byzantine_ledger.rs into separate files
