--- consensus/service/src/byzanine_ledger/mod.rs
@@ -1,9 +1,0 @@
-// Copyright (c) 2018-2020 MobileCoin Inc.
-
-//! A Federated, Byzantine Fault-Tolerant Ledger.
-//!
-//! Orchestrates running single-slot consensus, or performing ledger sync with peers.
-
-mod ledger_sync_state;
-mod task_message;
-mod worker;
