--- consensus/service/src/byzanine_ledger/worker.rs
@@ -1,1 +1,0 @@
-// Copyright (c) 2018-2020 MobileCoin Inc.
